Pass in bundle to validate against
The bundles are not loaded at runtime when we use the gem for
validation. In order to use the bundle for validation we need to pass in
a bundle. We tried in the initializer where they are defined, but the
database connection is not available there so this seems like the
easiest way to accomplish this.